wsd: move remote config classes to own home
This is a pure refactoring, without any code
change. It moves over 900 lines of code
from COOLWSD.cpp into RemoteConfig.hpp to
reduce the over-grown size of the former.
